Adjust leaflet control elements when map controls get rerendered
